+++ packages/lit-element/src/test/lit-element_test.ts
@@ -20,10 +20,7 @@ import {
   Part,
   nothing,
 } from '../lit-element.js';
-import {
-  directive,
-  DisconnectableDirective,
-} from 'lit-html/disconnectable-directive.js';
+import {directive, AsyncDirective} from 'lit-html/async-directive.js';
 import {
   canTestLitElement,
   generateElementName,
@@ -353,7 +350,7 @@ import {createRef, ref} from 'lit-html/directives/ref.js';
       const log: string[] = [];
 
       const d = directive(
-        class extends DisconnectableDirective {
+        class extends AsyncDirective {
           id!: unknown;
           render(id: unknown) {
             log.push(`render-${id}`);
@@ -607,7 +604,7 @@ import {createRef, ref} from 'lit-html/directives/ref.js';
   test('directive as controller can be added/removed via connect/disconnect', async () => {
     const log: string[] = [];
     const controllerDirective = directive(
-      class extends DisconnectableDirective {
+      class extends AsyncDirective {
         part?: Part;
         host?: Host;